feat(platform): deliver realtime localization and runtime integration updates
Unify backend and frontend runtime behavior for localization, floorplan presence, and role dashboards while expanding MCP and agent-runtime integration. Improve live map responsiveness, readiness repair flows, and workspace-scoped operational paths to reduce stale state and inconsistent room mapping.

+++ firmware/M5StickCPlus2/src/ui/DisplayManager.h
@@ -17,7 +17,7 @@ public:
     
     // UI Elements
     void drawMenu(const char* title, const char* items[], int itemCount, int selectedIndex, bool showGuideFooter = true);
-    void drawMessage(const char* title, const char* msg, uint16_t color = COLOR_TEXT);
+    void drawMessage(const char* title, const char* msg, uint32_t color = COLOR_TEXT);
     void drawKeyboard(const char* title, char* buffer, int maxLength); // Logic handled in scene, this draws
     
     // Helper
